[ADD] Filter refactor
