+++ src/pages/Home/index.js
@@ -12,45 +12,16 @@ import "../Home/Home.css";
 function Home() {
   // fetch Data
   const [films, setFilms] = useState([]);
+  // Filter toggle state
+  const [toggleFilter, setToggleFilter] = useState(false);
 
   useEffect(async () => {
     const response = await axios.get("https://ghibliapi.herokuapp.com/films");
     setFilms(response.data);
   }, []);
-
-  // Filter state
-  const [adjust, setAdjust] = useState("");
-  const [categories, setCategories] = useState("");
-  const [toggleFilter, setToggleFilter] = useState(false);
-
+  // Filter toggle function
   const handleToggle = () => {
     setToggleFilter(!toggleFilter);
-  };
-  // Filter engine
-  /*  Documentation
-    - sortArray calls two arguments category and order
-    - initialize categories variable that is an object
-    - every propery is an api property as well
-    - then another variable that storage an array property
-    - initialize a variable that executes the sort
-    - this is a validation, if order(argument) is equal to "ascendant"
-    - then complement or change the content in films with the sort order
-    - else executes same sort with the reverse order
-    - at last update the films state with the "sorted" results
-  */
-  const sortArray = (category, order) => {
-    const categories = {
-      running_time: "running_time",
-      release_date: "release_date",
-      rt_score: "rt_score",
-    };
-    const sortProperty = categories[category];
-    const sorted =
-      order === "ascendant"
-        ? [...films].sort((a, b) => b[sortProperty] - a[sortProperty])
-        : [...films].sort((a, b) => a[sortProperty] - b[sortProperty]);
-
-    setFilms(sorted);
   };
   return (
     <>
@@ -65,17 +36,7 @@ function Home() {
         <SearchEngine films={films} />
         <span onClick={handleToggle} className="filter-icon" />
       </div>
-      {!!toggleFilter && (
-        <Filter
-          sortArray={sortArray}
-          adjust={adjust}
-          setAdjust={setAdjust}
-          categories={categories}
-          setCategories={setCategories}
-          films={films}
-          setFilms={setFilms}
-        />
-      )}
+      {!!toggleFilter && <Filter films={films} setFilms={setFilms} />}
       <div className="film-cards-container">
         {films.map((item) => (
           <div style={{ margin: "10px" }} key={item.id}>
